correct datagouv search tests
